+++ src/main/java/com/global/api/network/entities/nts/NtsDataCollectRequestBuilder.java
@@ -47,8 +47,9 @@ public class NtsDataCollectRequestBuilder implements INtsRequestMessage {
             NtsUtils.log("Entry Method", NTSEntryMethod.MagneticStripeWithoutTrackDataAttended);
         } else if (paymentMethod instanceof GiftCard) {
             GiftCard card = (GiftCard) paymentMethod;
-            request.addRange(card.getEntryMethod(), 1);
-            NtsUtils.log("Entry Method", card.getEntryMethod());
+            NTSEntryMethod entryMethod = NtsUtils.isAttendedOrUnattendedEntryMethod(card.getEntryMethod(), card.getTrackNumber(), ntsObjectParam.getNtsAcceptorConfig().getOperatingEnvironment());
+            request.addRange(entryMethod.getValue(), 1);
+            NtsUtils.log("Entry Method", entryMethod);
         }
 
         // Card Type